requirements.txt
riskfolio-lib>=7.2.1
yfinance>=1.2.0
pandas>=2.2.0
numpy>=1.26.0
matplotlib>=3.9.0
scipy>=1.11.0
scikit-learn>=1.3.0
statsmodels>=0.14.0
cvxpy>=1.6.0
xlsxwriter>=3.2.0
openpyxl>=3.1.0
ollama>=0.6.1
